Virtual-Polymer/VFS/ROMP_and_polyimide/Code_to_generate_structures/pyproject.toml
[tool.poetry]
name = "polyverse"
version = "0.2.0"
description = ""
authors = ["Rishi Gurnani <rgurnani96@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pandas = ">1.5.3"
tqdm = "^4.66.1"
matplotlib = "^3.8.0"
requests = "^2.32.0"
dask = "^2023.9.2"
names = "^0.3.0"
CairoSVG = "^2.7.1"
openpyxl = "^3.1.2"
selenium = "^4.12.0"
more-itertools = "^10.1.0"
rdkit-pypi = ">=2021.9.3"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
black = "^24.3.0"
ipython = "^8.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
